Don't set fan speed if stored fan level matches
I could be wrong but it's my understanding that in manual mode, it doesn't need reminders:
the EC will keep the speed manually set

+++ FanCtrlCommon/FanCtrlInterface.cs
@@ -20,8 +20,8 @@ namespace FanCtrlCommon
         [DataMember]
         public uint SystemTemperature { get; private set; }
         [DataMember]
-        public sbyte FanLevel { get; private set; }
-        public FanCtrlData(uint temp, sbyte fan)
+        public uint FanLevel { get; private set; }
+        public FanCtrlData(uint temp, uint fan)
         {
             SystemTemperature = temp;
             FanLevel = fan;
